Edited multiplayer requirement.
The rubric is for the whole class, not just our project and we know that
one of the projects isn't going to have a multiplayer component.  However,
I feel that the social aspect of this course is important, and so I think
we should require that non-multiplayer games at least have some type of
score leader board or other social component.